+++ src/app/modules/user-management/component/user/user.component.ts
@@ -6,6 +6,7 @@ import { AddUserComponent } from '../add-user/add-user.component';
 import { PageEvent } from '@angular/material/paginator';
 import { ToastService } from 'src/app/shared/services/toast.service';
 import { User } from 'src/app/interface';
+import { ConfirmDialogService } from 'src/app/shared/services/confirm-dialog.service';
 
 @Component({
   selector: 'app-user',
@@ -20,7 +21,7 @@ export class UserComponent {
   page: number = 1;
   limit: number = 20;
   totalItems: any
-  defaultImg:string = "./../../../../../assets/default.jpeg"
+  defaultImg: string = "./../../../../../assets/default.jpeg"
 
   trackBy: TrackByFunction<any> = (index: number, item: User) => item.id ?? index;
 
@@ -28,7 +29,8 @@ export class UserComponent {
     private userService: UserService,
     private sharedService: SharedService,
     private dialog: MatDialog,
-    public toastService: ToastService
+    public toastService: ToastService,
+    private confirmService: ConfirmDialogService
   ) { }
 
   ngOnInit() {
@@ -73,14 +75,21 @@ export class UserComponent {
   }
 
   deleteUser(user: User) {
-    this.userService.deleteUser(user?.id).subscribe((res: User) => {
-      if (res) {
-        this.toastService.add('User successfully deleted.', 2000, "success");
-        this.getUsers()
-      }
-    }, (error) => {
-      this.toastService.add('Something went wrong!', 2000, "error");
-    });
+
+    this.confirmService
+      .confirm('Confirm Deletion', `Are you sure you want to delete User?`)
+      .subscribe((confirmed) => {
+        if (confirmed) {
+          this.userService.deleteUser(user?.id).subscribe((res: User) => {
+            if (res) {
+              this.toastService.add('User successfully deleted.', 2000, "success");
+              this.getUsers()
+            }
+          }, (error) => {
+            this.toastService.add('Something went wrong!', 2000, "error");
+          });
+        }
+      });
   }
 
   onFilterChanged(filtered: any[]): void {
